voice2text/requirements.txt
torch==2.0.1
torchaudio==2.0.2
openai-whisper==20231117
flask==2.3.3
pydub==0.25.1
numpy==1.24.4